requirements.txt
alembic==1.8.1
click==8.1.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-Cors==3.0.10
Flask-Login==0.6.2
flask-marshmallow==0.14.0
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
gunicorn==20.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
marshmallow==3.19.0
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-dotenv==0.21.0
six==1.15.0
SQLAlchemy==1.4.44
Werkzeug==2.2.2
WTForms==3.0.1
